+++ changelog.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Laravel Decomposer will be documented in this file
 
+## v1.0 (01-03-2017)
+- Added search,sorting & pagination on the table containing name & versions of installed packages and their dependencies.
+- Decomposes your Laravel & Server environment and gives all the required details & configurations about both on the same page.
+- Generate the system report of your Laravel & server environment on a single click. This report can be very useful in troubleshooting any error. Why & what about the system report is documented [here](https://github.com/lubusIN/laravel-decomposer/blob/master/report.md)
+
 ## v0.1.4 (26-02-2017)
 - Added check for no dependencies